requirements.txt
streamlit
google-generativeai
importlib-metadata
